Move bootutil_img_hash into utility file
This function will be shared by the current image verification and the
X509 code, so move it into a separate file.
