+++ boot/zephyr/CMakeLists.txt
@@ -133,6 +133,7 @@ zephyr_library_sources(
   ${BOOT_DIR}/bootutil/src/swap_misc.c
   ${BOOT_DIR}/bootutil/src/swap_scratch.c
   ${BOOT_DIR}/bootutil/src/swap_move.c
+  ${BOOT_DIR}/bootutil/src/image_util.c
   ${BOOT_DIR}/bootutil/src/caps.c
   )
 endif()
